Move `PowerManagingActor` to the `microgrid` module
