+++ src/frequenz/sdk/microgrid/_power_managing/_base_classes.py
