+++ src/frequenz/sdk/microgrid/_power_managing/_matryoshka.py
