+++ src/frequenz/sdk/microgrid/_power_managing/_sorted_set.py
