+++ src/frequenz/sdk/timeseries/battery_pool/_battery_pool_reference_store.py
@@ -16,10 +16,10 @@ from frequenz.client.microgrid import ComponentCategory
 from ..._internal._asyncio import cancel_and_await
 from ..._internal._channels import ChannelRegistry, ReceiverFetcher
 from ...actor._data_sourcing._component_metric_request import ComponentMetricRequest
-from ...actor._power_managing._base_classes import Proposal, ReportRequest
 from ...actor.power_distributing import Result
 from ...actor.power_distributing._component_status import ComponentPoolStatus
 from ...microgrid import connection_manager
+from ...microgrid._power_managing._base_classes import Proposal, ReportRequest
 from ..formula_engine._formula_engine_pool import FormulaEnginePool
 from ._methods import MetricAggregator
