+++ tests/actor/_power_managing/test_report.py
@@ -3,7 +3,7 @@
 
 """Tests for methods provided by the PowerManager's reports."""
 
-from frequenz.sdk.actor._power_managing import _Report
+from frequenz.sdk.microgrid._power_managing import _Report
 from frequenz.sdk.timeseries import Bounds, Power
